# Move refcounting to an array storage type