--- a/test/pointer.jl
+++ b/test/pointer.jl
@@ -41,7 +41,7 @@ ccall(:clock, Nothing, (Ptr{Int},), a)
 @test_throws Exception ccall(:clock, Nothing, (CuPtr{Int},), a)
 ccall(:clock, Nothing, (PtrOrCuPtr{Int},), a)
 
-b = CuArray{eltype(a), ndims(a)}(CU_NULL, size(a))
+b = CuArray{eltype(a), ndims(a)}(CUDA.ArrayStorage(Mem.DeviceBuffer(CU_NULL, 0), 0), size(a))
 ccall(:clock, Nothing, (CuPtr{Int},), b)
 @test_throws Exception ccall(:clock, Nothing, (Ptr{Int},), b)
 ccall(:clock, Nothing, (PtrOrCuPtr{Int},), b)