Simplifies the allocator: Instead of refcounting using a global dict, we now keep track using an atomic struct field. This avoids having to take a lock in the allocation path, improving performance.

Before:

```
julia> @benchmark CuArray{Int}(undef, 1024)
BenchmarkTools.Trial: 
  memory estimate:  208 bytes
  allocs estimate:  7
  --------------
  minimum time:     852.206 ns (0.00% GC)
  median time:      987.778 ns (0.00% GC)
  mean time:        1.979 μs (5.39% GC)
  maximum time:     3.789 ms (9.53% GC)
  --------------
  samples:          10000
  evals/sample:     63
```

After:
```
julia> @benchmark CuArray{Int}(undef, 1024)
BenchmarkTools.Trial: 
  memory estimate:  128 bytes
  allocs estimate:  4
  --------------
  minimum time:     574.550 ns (0.00% GC)
  median time:      656.050 ns (0.00% GC)
  mean time:        1.604 μs (4.02% GC)
  maximum time:     1.825 ms (6.15% GC)
  --------------
  samples:          10000
  evals/sample:     180
```