@jonathan-laurent 2021-07-01
Looks great! I'll test this on AlphaZero.jl as soon as https://github.com/JuliaGPU/CUDA.jl/issues/1025 is elucidated.